ID-47: Add route to allow new account holders to see their email verification token validity
This will allow donors to see that a token is valid for up to 8 hours
after it was generated. When we implement the route to actually *use*
the token to set a password we could add another 5 minutes of validity
there in case they take a time to choose a password.